Basic matching for function and macro definitions
Updated the matching for functions and for macros so that the
function name/macro name is highlighted correctly. Also updated the
snippets for both functions and macros so that they create function
names by default. Without providing a name it highlights the `do` as the
function name which looks odd.